modified:   anag_stats_/experiment.js 	modified:   anag_stats_/experiment/anagramer/jspysch/core_js/plugin-gram.js
Progress 7-15, resolved the outstanding timline
problems that arose out of trying to do an
on the fly randomization.
For now am settling HTML window errors in the
plugin-gram.js file which was orignally part of
a cloze experiment, was mod-ed and then further
mod-ed to be a part of the anagramer experiment.

The current issue now that the  TV array which
contains the params for the gram plugin
isn't being called correctly in the plugin class
part of the script. I've double starred the line I
have been trying to alter.